Fix `tools.jar` missing when using forked compilation
... and when the target JDK is _not_ the JDK which runs Gradle.
Without this jar, the classes from the compiler API are missing.
